+++ bake/traces/provider.rb
@@ -1,0 +1,15 @@
+# List all available trace providers.
+def list
+	available = {}
+	
+	Gem.loaded_specs.each do |name, spec|
+		spec.require_paths.each do |require_path|
+			root = File.expand_path(require_path, spec.full_gem_path)
+			Dir.glob("traces/provider/**/*.rb", base: root).each do |path|
+				(available[name] ||= []) << path
+			end
+		end
+	end
+
+	return available
+end